Day 14 Part 02.

+++ AdventOfCode/AdventOfCode/Solvers/Day14/Chemical.cs
@@ -1,9 +1,9 @@
 namespace AdventOfCode.Solvers.Day14 {
   public class Chemical {
-    public int Units { get; set; }
+    public long Units { get; set; }
     public string Name { get; set; }
 
-    public Chemical(int units, string name) {
+    public Chemical(long units, string name) {
       Units = units;
       Name = name;
     }
